backend/requirements.txt
alembic==1.14.0
Authlib==1.4.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
eventlet==0.38.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-PyMongo==2.3.0
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
oauthlib==2.1.0
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pymango==0.1.1
pymongo==4.10.1
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.12.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.1.0
simple-websocket==1.1.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
